+++ zojae031/Project/app/src/main/java/study/architecture/BaseFragment.kt
@@ -1,0 +1,21 @@
+package study.architecture
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.databinding.DataBindingUtil
+import androidx.databinding.ViewDataBinding
+import androidx.fragment.app.Fragment
+
+open class BaseFragment<B : ViewDataBinding>(private val layoutId: Int) : Fragment() {
+    protected lateinit var binding: B
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = DataBindingUtil.inflate(inflater, layoutId, container, false)
+        return binding.root
+    }
+}
